E-18: extract ImportedMemberRefFactory from ReflectionMetadataEmitter (#1361)
Move the imported/BCL member-and-type reference factory band out of
ReflectionMetadataEmitter into a new internal sealed class
ImportedMemberRefFactory (src/Core/CodeAnalysis/Emit/ImportedMemberRefFactory.cs).
RME 10,104 -> 8,101 lines; new file 2,231 lines.

Band moved (verbatim bodies, mechanical rewrites only):
- Type refs: GetElementTypeToken, GetTypeOfToken, GetAssemblyReference,
  GetSystemRuntimeAssemblyReference, IsCoreLibBaseType, GetTypeReference,
  GetTypeHandleForMember.
- Method/ctor/field refs: GetMethodReference, GetMethodEntityHandle (x4),
  GetCtorReference (x2), GetFieldReference, GetFieldReferenceOnConstructedGeneric,
  GetCtorReferenceOnConstructedGeneric, GetDelegateCtorReference, and their
  symbolic-container helpers (TryCreate*ForConstructedSymbolicContainer,
  TryFindImplementedInterfaceInstantiation, SameOpenTypeDefinition,
  TryNormalizeToSymbolicContainer, GetErasedObjectArgs, ChooseErased*).
- Open-definition / TypeBuilder-safe reflection helpers: GetOpenMethod,
  GetOpenCtor, SameMetadataDefinition, ContainsTypeBuilderGenericArgument,
  ResolveTypeBuilderConstructedGeneric{Method,Ctor,Field},
  GetTypeBuilderSafePropertyAccessor, ResolveMethodOnOpenDefinition,
  ResolvePropertyOnOpenDefinition, ResolveCtorOnOpenDefinition.
- Nullable ref family: GetNullableCtorMemberRefForOpenTypeParameter,
  GetNullableCtorMemberRefForUserEnum, GetNullableCtorMemberRefForUserValueType,
  GetNullableGetValueMemberRefForUserValueType,
  GetNullableGetHasValueMemberRefForUserValueType.
- Tuple/map ref families: GetTupleTypeSpec, GetTupleFieldReference,
  GetTupleCtorReference, GetMapTypeSpec, GetMapCtorReference, GetMapSetItemReference.
- Reified Func/Action delegate MemberRef producers: GetFunctionDelegateTypeSpec,
  GetConstructedDelegateCtorRef, GetFunctionDelegateCtorRef,
  GetFunctionDelegateInvokeRef, TryGetSymbolicAsyncDelegateCtorRef.

Boundary helpers left on RME (E-19 material / general predicates, reached via
outer. back-ref from the factory): GetUserStructTypeSpec, GetUserInterfaceTypeSpec,
IsUserGenericTypeReference, IsUserGenericInterfaceReference, ArgIsSymbolicUserDefined,
FunctionTypeNeedsSymbolicDelegate, stateMachines plans, FindImportedMethod,
ResolveCoreType, BuildUnmanagedConstraintTypeSpec.

State-location findings: this was the near-pure move the plan flagged. Every
ref-cache dictionary the band writes already lives on MetadataTokenCache (E-2)
and is reached via cache.X (AssemblyRefs, SystemRuntimeAssemblyRef, TypeRefs,
TypeSpecs, MethodRefs, MethodSpecs, MethodSpecsWithSymbolArgs, CtorRefs,
CtorRefsWithSymbolArgs, NullableOpenCtorMemberRefs, NullableUserEnumCtorMemberRefs,
NullableUserStructCtorMemberRefs, NullableUserValueTypeGetValueMemberRefs,
NullableUserValueTypeGetHasValueMemberRefs, FieldRefs, StructTypeDefs,
EnumTypeDefs, InterfaceTypeDefs) — NO metadata state relocated. The only fields
that moved are the three reified-delegate-shape caches
(functionDelegateTypeSpecCache / functionDelegateCtorRefCache /
functionDelegateInvokeRefCache), which were RME privates consumed solely by the
delegate MemberRef producers; they are now private fields of the factory.

Design: ctor (ReflectionMetadataEmitter outer) back-ref with direct convenience
fields for emitCtx / cache / remaps / signatures. Signature encoding routes
through signatures.EncodeTypeSymbol / EncodeClrType / EncodeReturnClr /
EncodeFunctionTypeSymbol (the E-17 SignatureEncoder). E-17's temporary
outer.GetTypeReference coupling still resolves via the root forwarder (no
SignatureEncoder change now; E-21 repoints). memberRefs is constructed alongside
signatures in the ctor (order verified: SignatureEncoder's ctor reads only
emitCtx/cache/remaps, and all cross-calls go through outer forwarders at
call-time, so no earlier use).

Forwarders for E-21 (30 one-line root forwarders kept so untouched call sites
compile): GetElementTypeToken, GetTypeOfToken, GetTypeReference,
GetTypeHandleForMember, GetTypeBuilderSafePropertyAccessor (static),
GetMethodReference, GetMethodEntityHandle (x4), GetCtorReference (x2),
GetNullableCtorMemberRefForOpenTypeParameter, GetNullableCtorMemberRefForUserEnum,
GetNullableCtorMemberRefForUserValueType, GetNullableGetValueMemberRefForUserValueType,
GetNullableGetHasValueMemberRefForUserValueType, GetFieldReference,
GetFieldReferenceOnConstructedGeneric, GetCtorReferenceOnConstructedGeneric,
GetDelegateCtorReference, GetTupleFieldReference, GetTupleCtorReference,
GetMapCtorReference, GetMapSetItemReference, GetFunctionDelegateTypeSpec,
GetConstructedDelegateCtorRef, GetFunctionDelegateCtorRef,
GetFunctionDelegateInvokeRef, TryGetSymbolicAsyncDelegateCtorRef.

Widenings: ResolveMethodOnOpenDefinition, ResolvePropertyOnOpenDefinition,
TryNormalizeToSymbolicContainer, SameMetadataDefinition widened
private static -> internal static on the factory (they are shared with the
root-retained TryGetSymbolicSubstituted{Property,InstanceMethod}Return query
helpers, whose three call sites are repointed to ImportedMemberRefFactory.X).

Verification: dotnet build GSharp.sln -c Release -graph -> 0 warnings, 0 errors.
PE byte-diff: all 119/119 samples byte-identical vs baseline (3 GsharpExtensions*
skipped).
